+++ DesignToolbox/DesignToolbox/Pages/Components/Controls/TextInput/TextInputConfigurationView.swift
@@ -125,6 +125,14 @@ final class TextInputConfigurationModel: ComponentConfiguration {
         didSet { updateCode() }
     }
 
+    @Published var progressVariant: CircularProgressIndicatorConfigurationModel.Variant {
+        didSet { updateCode() }
+    }
+
+    @Published var progressValue: Double {
+        didSet { updateCode() }
+    }
+
     // MARK: Initializer
 
     override init() {
@@ -146,6 +154,10 @@ final class TextInputConfigurationModel: ComponentConfiguration {
         constrainedMaxWidth = false
         status = .enabled
         textMode = .raw
+
+        progressVariant = .indeterminate
+        progressValue = 0.0
+
         super.init()
     }
 
@@ -171,7 +183,7 @@ final class TextInputConfigurationModel: ComponentConfiguration {
         case .richError:
             .richError(message: richErrorText)
         case .loading:
-            .loading
+            .loading(progress: progressVariant == .indeterminate ? nil : progressValue)
         case .readOnly:
             .readOnly
         case .disabled:
@@ -282,7 +294,11 @@ final class TextInputConfigurationModel: ComponentConfiguration {
         case .richError:
             ", status: .richError(message: yourAttributedString)"
         case .loading:
-            ", status: .loading"
+            if progressVariant == .indeterminate {
+                ", status: .loading"
+            } else {
+                ", status: .loading(progress: \(String(format: "%.2f", progressValue)))"
+            }
         case .readOnly:
             ", status: .readOnly"
         case .disabled:
@@ -331,6 +347,16 @@ struct TextInputConfigurationView: View {
                 OUDSChipPicker(title: "app_components_common_status_tech",
                                selection: $configurationModel.status,
                                chips: TextInputStatus.chips)
+
+                if configurationModel.status == .loading {
+                    OUDSChipPicker(title: "app_components_progressIndicator_variant_tech",
+                                   selection: $configurationModel.progressVariant,
+                                   chips: CircularProgressIndicatorConfigurationModel.Variant.chips)
+
+                    if configurationModel.progressVariant == .determinate {
+                        ProgressControl(progress: $configurationModel.progressValue)
+                    }
+                }
 
                 if configurationModel.status != .error, configurationModel.status != .richError {
                     OUDSChipPicker(title: "app_components_textMode_tech",